Expand mapping space to allow all alphabetics
Expand nodes to cater for A-Z, a-z, 0-9, _ and ' chars to
allow full dictionary of more unusual words.  This means the
offsets can be 32 bit in size, so remove the 24/32 bit squashed
indexing.  Make mapping from chars to index into a node more
explicit.

Make root node of dictionary start at index 0 and first new
word start at index 1.  This allows us to get around a zero'th
index offset bug that always made the first word in the
dictionary unfindable.

Signed-off-by: Colin Ian King <colin.king@canonical.com>