Fix detection of printk-like calls with whitespaces
Unexpectedly, there are function calls in the Linux kernel that
violate the coding style of no whitespaces between the function
name and opens parentheses.

In the latest stable release (currently 5.11.6), there are 1468
occurrences of the type (about 0.47% of the total calls to
printk-like functions), most of them with line breaks to avoid
exceeding the 80 columns, as in:

drivers/scsi/wd33c93.c:1398
if (!( asr & ASR_INT)) {
	printk
		    ("wd33c93: Reselected without IDENTIFYn");
	lun = 0;
} else {

Due to kernelscan assumes there are no spaces, these calls are
not identified in a normal kernelscan invocation, with no extra
options.

Fixes this by checking that the current token is a whitespace,
and if so, parses it.

Signed-off-by: Davidson Francis <davidsondfgl@gmail.com>